OfferDetailComp
